Move attestation generation to attester
